requirements.txt
cellxgene-census[experimental]==1.4.0
lightning==2.0.6
scvi-tools==1.0.2
fastapi==0.100.1
click==7.1.2